Add initial setup to cody web lib build
