+++ pnpm-workspace.yaml
@@ -1,7 +1,7 @@
 packages:
-  - agent
-  - agent/src/__tests__/*
-  - cli
-  - lib/*
-  - vscode
-  - web
+  - 'agent'
+  - 'agent/src/__tests__/*'
+  - 'cli'
+  - 'lib/*'
+  - 'vscode'
+  - 'web'
