+++ vscode/webviews/App.module.css
@@ -17,7 +17,6 @@
     padding: 1rem;
     color: var(--vscode-input-foreground);
     background-color: var(--vscode-inputValidation-errorBackground);
-    align-items: center;
     min-height: 2rem;
     position: relative;
     overflow: auto;
