+++ web/lib/agent/shims/env-paths.ts
