+++ web/lib/agent/shims/inline-completion-item-provider.ts
