+++ web/lib/agent/shims/stream.ts
