+++ web/lib/cody-web-chat.css
@@ -11,16 +11,15 @@
     --vscode-editor-font-family: 'Monaco', 'Consolas', monospace;
 }
 
-body {
-    margin: 0;
-
+[data-cody-web-chat] {
+    height: 100%;
     background-color: var(--vscode-editor-background);
     font-size: var(--vscode-font-size);
     font-family: var(--vscode-font-family);
     color: var(--vscode-editor-foreground);
 }
 
-a {
+[data-cody-web-chat] a {
     color: var(--vscode-textLink-foreground);
     text-decoration: none;
 }
