+++ web/lib/globals.d.ts
@@ -1,0 +1,4 @@
+declare module '*.module.css' {
+    const classes: { readonly [key: string]: string }
+    export default classes
+}
