+++ web/lib/index.ts
@@ -1,0 +1,3 @@
+
+export { CodyWebChat } from './Chat'
+export { type CodyWebChatProps } from './Chat'
