+++ web/package.json
@@ -1,23 +1,25 @@
 {
-  "private": true,
   "name": "@sourcegraph/cody-web",
   "version": "0.0.1",
   "description": "Cody standalone web app",
   "license": "Apache-2.0",
+  "main": "dist/index.js",
+  "types": "dist/lib/index.d.ts",
   "repository": {
     "type": "git",
     "url": "https://github.com/sourcegraph/cody",
     "directory": "web"
   },
-  "main": "src/index.ts",
   "sideEffects": false,
   "scripts": {
     "dev": "vite --mode development",
-    "build": "tsc --build && vite build --mode production",
+    "build": "tsc --build tsconfig.json --emitDeclarationOnly && vite build --mode production",
     "test": "vitest"
   },
   "dependencies": {
+    "@sourcegraph/cody-agent": "workspace:*",
     "@sourcegraph/cody-shared": "workspace:*",
+    "@sourcegraph/vscode-cody": "workspace:*",
     "@vitest/web-worker": "^1.4.0",
     "@vscode/codicons": "^0.0.35",
     "buffer": "^6.0.3",
@@ -27,10 +29,16 @@
     "util": "^0.12.5",
     "vscode-uri": "^3.0.7"
   },
+  "peerDependencies": {
+    "react": "18.2.0",
+    "react-dom": "18.2.0"
+  },
   "devDependencies": {
+    "@types/highlightjs": "^9.12.6",
+    "@vitejs/plugin-react": "^4.2.1",
     "@vitejs/plugin-react-swc": "^3.6.0",
     "postcss": "^8.4.38",
     "tailwind-merge": "^2.3.0",
-    "tailwindcss": "^3.4.3"
+    "vite-plugin-dts": "^3.9.1"
   }
 }
