--- web/src/sample.test.ts
@@ -1,3 +1,0 @@
-import { expect, test } from 'vitest'
-
-test('should pass', () => expect(1).toBe(1))
